pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hcpya_preprocess"
version = "0.1.0"
description = "HCP-Youth BOLD Pre/Post-Processing"
authors = [
    {name = "Michael Demidenko", email = "demidenko.michael@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy==1.23.5",        # Stable version that avoids breaking changes with other packages
    "matplotlib==3.5.3",     # Compatible with Seaborn 0.11.2 and PtitPrince
    "nibabel",
    "nilearn",
    "nipype",
    "pandas==1.3.5",        # Works with Seaborn 0.11.2
    "seaborn==0.11.2",      # Compatible with PtitPrince
    "ptitprince",
    "pyrelimri",
    "svgwrite",
    "jupyter",
    "notebook",
    "nbformat",
    "ipython",
    "pybids"
]


[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
packages = ["scripts"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
